Se agregó el modelo Producto y pruebas unitarias

+++ productos/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class Producto(models.Model):
     nombre = models.CharField(max_length=100)
     descripcion = models.TextField()
-    precio = models.IntegerField()
+    precio = models.DecimalField(max_digits=10, decimal_places=2)
     stock = models.IntegerField()
 
     def __str__(self):
